% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{boot_cluster}
\alias{boot_cluster}
\title{Resample clustered data}
\usage{
boot_cluster(dat, cluster.vars, out.folder, i)
}
\arguments{
\item{dat}{The original data to be bootstrapped (does not contain a variable
named "cluster").}

\item{cluster.vars}{An unquoted vector of column names used to construct a
cluster. An unquoted vector of column names.}

\item{out.folder}{A quoted string that gives the path to an existing folder
where bootstrapped data objects should be written.}

\item{i}{a counter that will be written to the name of each bootstrapped data
(e.g bootdat_1.rds, dat_2.rds)}
}
\value{
Writes an RDS file containing one bootstrapped data set.
}
\description{
This function defines the cluster variables, then resamples from the unique
clusters N times, where N is the number of unique clusters. Observations
within each cluster are also resampled.
}
