% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPestimates.R
\name{chapman}
\alias{chapman}
\title{Chapman Estimator for closed populations}
\usage{
chapman(r, n, m)
}
\arguments{
\item{r}{the number of individuals marked in the initial sample.}

\item{n}{the total number of individuals (marked and unmarked) encountered in
the second sample.}

\item{m}{the number of marked individuals encountered in the second sample}
}
\value{
A list containing:

\item{N.hat}{the abundance estimate}

\item{SE}{the standard error of the abundance estimate}
}
\description{
Uses Chapman's modified version of the Lincoln-Petersen estimator for a
two-sample closed population abundance estimator.
}
\examples{
\dontrun{
 # Use simulation code to generate data

 LP.sim(N=5E4, mark=300, recap=0.2,sample.fixed=TRUE)->sim.dat


 chapman(r=sim.dat$r,
          n=sim.dat$n,
          m=sim.dat$m)
}
}
