% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties.mo}
\alias{counties.mo}
\title{Missouri county boundaries}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 115 rows and 7 columns.
}
\usage{
counties.mo
}
\description{
A sf object containing  polygons representing the border of each county
 of Missouri. The boundary was derived by dissolving the county boundaries
 from the tigris, county data layer. It is in NAD83 UTM 15N (meters)
 coordinate system. Per the tigris package, it is 1:500k resolution (default)
}
\examples{
\dontrun{
tigris::counties(state='MO')\%>\%
sf::st_transform(., crs=26915)\%>\% # reproject to NAD83 UTM 15N
 select(STATEFP,
        COUNTYFP,GEOID,
        NAME,
        NAMELSAD,
        LSAD)->counties.mo
}
}
\keyword{datasets}
