% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_functions.R
\name{est_accelerate}
\alias{est_accelerate}
\title{Jacknife procedure to estimate the acceleration constant}
\usage{
est_accelerate(OG.dat, dat.list, est.fxn)
}
\arguments{
\item{OG.dat}{The orignal data set, a data frame.}

\item{dat.list}{A list of Jacknifed data sets. Generated internally in \code{bca_jacknife}}

\item{est.fxn}{A user-supplied function that performs the desired operation on the data.
Typically a call to \code{lm}, \code{lmer}, etc. or any object that can then be coerced
into a data.frame similar to that returned by \code{broom::tidy} (or
\code{broom.mixed::tidy}).

Must return a data.frame with columns "term" and "estimate", representing
the parameter name, and it's estimate.}
}
\value{
A list comprised of 2 data frames. The first data frame has the first column
(term) indicating each parameter and the second column (a), the acceleration
constant for each term.
}
\description{
A function called internal in \code{bca_jacknife} to compute the acceleration
constant for each parameter of interest.
}
