% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{exploit_fish}
\alias{exploit_fish}
\title{Estimate fish exploitation rate}
\usage{
exploit_fish(release, recover, tag.loss, report.rate)
}
\arguments{
\item{release}{the number of tagged or marked fish releases}

\item{recover}{the number of tags recovered}

\item{tag.loss}{the probability a tag or other mark is lost}

\item{report.rate}{the probability a harvested, marked fish is reported}
}
\value{
the estimated exploitation rate (harvest probability)
}
\description{
Estimate fish exploitation rate
}
\examples{

exploit_fish(release=100,
             recover=10,
             tag.loss=0.1,
             report.rate=0.75)


}
