% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_params.R
\name{get_betadist}
\alias{get_betadist}
\title{Convert a probability and SD to beta distribution parameters}
\usage{
get_betadist(mu, sd)
}
\arguments{
\item{sd}{}
}
\value{
A named vector including alpha and beta parameters
}
\description{
Convert a probability and SD to beta distribution parameters
}
