% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPestimates.R
\name{LP.sim}
\alias{LP.sim}
\title{Lincoln-Petersen Simulation}
\usage{
LP.sim(N, mark, recap, sample.fixed = FALSE)
}
\arguments{
\item{N}{Simulated population size}

\item{mark}{the number of individuals marked or the probability an individual
is caught and marked during the initial sample period.}

\item{recap}{the probability an individual is encountered in the second
sample period}

\item{sample.fixed}{is the sample size fixed or not. If TRUE, \code{mark} is a
number >1 and represents a number of known marked individuals released, otherwise, it's a probability (0,1)}
}
\value{
A list containing:

\item{r}{the number of individuals marked in the first sample period}

\item{n}{the number of individuals in the second sample.}

\item{m}{the number of previously marked individuals in the second sample.}
}
\description{
Simulate data for a closed, 2-sample mark/recapture study
}
\examples{
\dontrun{
 LP.sim(N=5E4, mark=300, recap=0.2,sample.fixed=TRUE)

 LP.sim(N=5E4, mark=0.3 recap=0.2,sample.fixed=FALSE)
}
}
