% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcd.area.tab}
\alias{nlcd.area.tab}
\title{NLCD land cover by area for each county in MO odd numbered years 2001-2023.}
\format{
\describe{
   \item{cty}{county name}
   \item{yr}{4 digit year}
   \item{crop.sqkm}{area in km^2 of crop}
   \item{develop.sqkm}{area in km^2 of developed}
   \item{forest.sqkm}{area in km^2 of forest}
   \item{grass.sqkm}{area in km^2 of grass}
   \item{water.sqkm}{area in km^2 of water}
   \item{wetland.sqkm}{area in km^2 of wetland}
   \item{other.sqkm}{area in km^2 of all other landcover types present}
}
}
\usage{
nlcd.area.tab
}
\description{
A dataset derived of Annual NLCD data- files were obtained via the FedData
package (NLCD) and tigir package (county boundaries). NLCD data was reprojected
to NAD83 UTM 15 N before areal calculations.
}
\details{
See vignettes for examples on how to batch process

 Cropping performed with terra::crop

Area converted to square km based on a 30x30m cell size and summing all
all pixels with in a given class. No other processing occurred.
}
\examples{
\dontrun{
#pixel classes were condensed as follows:


terra::rast(nlcd)->Z
 mo.co.list[[county]]->Q
 terra::crop(Z,Q)\%>\%
 terra::freq()\%>\%
 filter(str_detect(value,
 'Water|Developed|Forest|Shrub|Grass|Hay|Crops|Wetlands'))\%>\%
 mutate(LC=case_when(str_detect(value,'Developed')~"develop",
                   str_detect(value,'Forest')~"forest",
                   str_detect(value,'Shrub|Grassland|Hay')~'grass',
                   str_detect(value,'Crops')~'crop',
                   str_detect(value,'Wetlands')~'wetland',
                   str_detect(value,'Water')~'water',
                   TRUE~as.character("other")
                   ))
                   }

}
\keyword{datasets}
