% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcd.pct.tab}
\alias{nlcd.pct.tab}
\title{#' NLCD land cover by percent of county area for each county in MO
odd numbered years 2001-2023.}
\format{

}
\usage{
nlcd.pct.tab
}
\description{
A dataset derived of Annual NLCD data- files were obtained via the FedData
package (NLCD) and tigir package (county boundaries). NLCD data was
reprojected to NAD83 UTM 15 N before areal calculations.
}
\details{
Cropping performed with terra::crop

Area converted to square km based on a 30x30m cell size and summing all
all pixels with in a given class. No other processing occured.
}
\examples{
\dontrun{
#pixel classes were condensed as follows:


terra::rast(nlcd)->Z
 mo.co.list[[county]]->Q
 terra::crop(Z,Q)\%>\%
 terra::freq()\%>\%
 filter(str_detect(value,
 'Water|Developed|Forest|Shrub|Grass|Hay|Crops|Wetlands'))\%>\%
 mutate(LC=case_when(str_detect(value,'Developed')~"develop",
                   str_detect(value,'Forest')~"forest",
                   str_detect(value,'Shrub|Grassland|Hay')~'grass',
                   str_detect(value,'Crops')~'crop',
                   str_detect(value,'Wetlands')~'wetland',
                   str_detect(value,'Water')~'water',
                   TRUE~as.character("other")
                   ))
                   }

}
\keyword{datasets}
