% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{resamp.fxn}
\alias{resamp.fxn}
\title{Resampling helper function}
\usage{
resamp.fxn(tmp.dat, tmp.name)
}
\arguments{
\item{tmp.dat}{the data to be bootstrapped, contains a "cluster" variable created in
boot_cluster}

\item{tmp.name}{The cluster value, used to subset data before resampling from observations
within the specified cluster.}
}
\value{
a data frame of resampled observations from within a single cluster
}
\description{
Function called internally in boot_cluster to resample observations within
a cluster. It uses a vector of resampled "clusters: (tmp.name), pulls just
observations from a given cluster in the original data (tmp.dat) and returns
a data frame of the resampled observations from a cluster. This is done
repeatedly within boot_cluster and the results are combined to produce a
single bootstrapped data set.
}
