% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.mo}
\alias{state.mo}
\title{Missouri state boundary}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
state.mo
}
\description{
A sf object containing a single polygon representing the border of Missouri.
The boundary was derived by dissolving the county boundaries from the tigris
county data layer. It is in NAD83 UTM 15N (meters) coordinate system.
}
\examples{
\dontrun{
tigris::counties(state='MO')\%>\%
sf::st_transform(., crs=26915)\%>\% # reproject to NAD83 UTM 15N
 select(STATEFP,
        COUNTYFP,GEOID,
        NAME,
        NAMELSAD,
        LSAD)->counties.mo

 counties.mo\%>\%
 group_by(STATEFP)\%>\%
 summarize()->state.mo
}


}
\keyword{datasets}
