% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_unmarked.R
\name{tidy.um}
\alias{tidy.um}
\title{tidy unmarked parameter estimates}
\usage{
tidy.um(X)
}
\arguments{
\item{X}{an unmarked fit object}
}
\value{
A data frame of parameter estimates similar to what broom::tidy()
does for lm/glm objects
}
\description{
tidy unmarked parameter estimates
}
