% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_unmarked.R
\docType{data}
\name{tidy.unm}
\alias{tidy.unm}
\title{tidy unmarked parameter estimates}
\format{
An object of class \code{function} of length 1.
}
\usage{
tidy.unm(X)

tidy.unm
}
\arguments{
\item{X}{an unmarked fit object}
}
\value{
A data frame of parameter estimates similar to what broom::tidy()
does for lm/glm objects
}
\description{
tidy unmarked parameter estimates
}
\examples{
No
}
\keyword{datasets}
