% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_unmarked.R
\name{tidy.unmarkedFit}
\alias{tidy.unmarkedFit}
\title{tidy unmarked parameter estimates}
\usage{
tidy.unmarkedFit(X)
}
\arguments{
\item{X}{an unmarked fit object}
}
\value{
A data frame of parameter estimates similar to what broom::tidy()
does for lm/glm objects
}
\description{
tidy unmarked parameter estimates
}
