% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_unmarked.R
\name{um.help}
\alias{um.help}
\title{Formatting for a single parameter}
\usage{
um.help(X)
}
\arguments{
\item{X}{would be one slot from within unmarkedFit@estimates list}
}
\value{
a data.frame of the proccess, parameter, SE and approximate Z and P
values (assuming normal distribution)
}
\description{
Formatting for a single parameter
}
