% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logexp.R
\name{logexp}
\alias{logexp}
\title{logistic exposure link function}
\usage{
logexp(exposure = 1)
}
\arguments{
\item{exposure}{The length of time. defaults to 1 unit.}
}
\description{
This is the logistic exposure link to use with glm's in R
It is sourced from Ben Bolker's website : https://rpubs.com/bbolker/logregexp
}
