% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_unmarked.R
\docType{data}
\name{um.help}
\alias{um.help}
\title{Formatting for a single parameter}
\format{
An object of class \code{function} of length 1.
}
\usage{
um.help(X)

um.help
}
\arguments{
\item{X}{would be one slot from within unmarkedFit@estimates list}
}
\value{
a data.frame of the proccess, parameter, SE and approximate Z and P
values (assuming normal distribution)
}
\description{
Formatting for a single parameter
}
\examples{
No
}
\keyword{datasets}
\keyword{internal}
